fixed mul mem_limit

+++ lab-12_exceptions/src/matrix.cpp
@@ -132,7 +132,7 @@ Matrix& Matrix::operator*=(const Matrix& m) {
     for(std::size_t i = 0; i < _rows; i++)
         for(std::size_t j = 0; j < m._cols; j++)
             for(std::size_t k = 0; k < _cols; k++)
-                mult._data[i][j] += _data[i][k]*m._data[k][j];    
+                mult._data[i][j] += (long long)_data[i][k]*m._data[k][j];    
 
     _cols = m._cols;
 
@@ -169,12 +169,7 @@ Matrix read_matrix_from_file(std::string filename){
     
     if(!(fs >> rows >> cols))
         throw MatrixException("LOAD: invalid file format.");
-
-    // if(!rows || !cols)
-    //     throw MatrixException("LOAD: invalid file format.");
-
-    // std::cout << rows << " " << cols << std::endl;
-    
+   
     int data[rows][cols];
     for(int i = 0; i < rows; i++)
         for(int j = 0; j < cols; j++){
